LayoutTests/discovery/discovery-base-api-expected.txt

Tests that getNetworkServices API is available and support all characters for valid service type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.getNetworkServices is defined.
PASS successfullyParsed is true

TEST COMPLETE
